Response > 고유번호는 int 대신 long 사용 처리

+++ src/main/java/study/msa/msadesigntaskmember/application/student/response/StudentRegistrationResponse.java
@@ -9,5 +9,5 @@ import lombok.*;
 @AllArgsConstructor
 public class StudentRegistrationResponse {
 
-    private int studentId; // 학생 고유번호
+    private long studentId; // 학생 고유번호
 }
